Обьединение веток

+++ src/main/java/ru/skypro/homework/service/ImageService.java
@@ -1,6 +1,7 @@
 package ru.skypro.homework.service;
 
 import org.springframework.web.multipart.MultipartFile;
+import ru.skypro.homework.model.Ad;
 import ru.skypro.homework.model.Image;
 import ru.skypro.homework.model.User;
 
@@ -10,5 +11,6 @@ public interface ImageService {
 
     void uploadImage(User user, MultipartFile imageFile) throws IOException;
     void uploadImage(Ad ad, MultipartFile imageFile) throws IOException;
-    Image findImage(long id);
+    Image findImage(int id);
+
 }
